+++ src/protocols/ftp.rs
@@ -24,6 +24,7 @@ impl FileTransferClient for FtpClient {
         _user: &str,
         _password: Option<&str>,
         _keyfile_path: Option<&str>,
+        _keyfile_passphrase: Option<&str>,
     ) -> Result<Self, FtpError>
     where
         Self: Sized,